requirements.txt
streamlit
gspread
google-auth
pandas
matplotlib
streamlit-autorefresh
